Fixed problem with newline characters
Fixed a problem the scanner in ConsoleInputOutput had with non-portable
use of the newline characters "\r\n." Changed to use
String.format("%n"), which should automatically insert the correct
newline characters depending on platform. Also added a message to the
save method in MainMenu when there is no activeSurvey/Test. Also
commented out an unused variable in Test for the time being.