Ability to modify MultipleChoice/TrueFalse
The ability to modify all aspects of MultipleChoice questions has been
added. The prompt can be changed (per the last commit), the choices can
be changed, and more choices can be added. Refactored MultipleChoice to
put the loop previously in addChoices() into a new loopChoices() method.
Changed the type of the Map used in MultipleChoice from TreeMap to
LinkedHashMap, to maintain insertion order instead of sorted order
(which makes more sense when more than 26 choices are added). Overrode
the added modifyChoices() method in TrueFalse to return, so now
TrueFalse can be totally modified.