Comparison of Responses; saving methodology
Changed the methodology of how Surveys/Tests are saved in the MainMenu.
It now keeps track of where the activeSurvey/Test_ is saved (under
survey/testPath_) and, when saving, prompts the user if they wish to
change the path to the saved Survey. Also added the required .equals()
and hashCode() methods to the Response class.