Added the ability to tabulate; variety of fixes
Can now tabulate Surveys and Tests. Implemented the tabulateAndDisplay()
method in Question class (and overrode it in the Essay hierarchy,
because Essays cannot be tabulated) to do so. Also fixed it so that now
multiple answers can be added to Essays. Also fixed the Matching class
to return the correct value in the getValidResponses() method; before it
created a Set out of the numbers of the leftColumn_, whereas it should
have created it out of the rightColumn_. Removed unnecessary (and
incorrect) overloads of .equals() and .hashCode() in the subclasses of
Response.