Fixed modification for Essays/ShortAnswers
Fixed it so that if the number of allowed Responses for a a ShortAnswer
is changed in a test using modifyQuestion(), the user will be
automatically prompted to enter the new correct answers; however, if
it's an Essay, the user will not be prompted for correct answers. In
addition, refactored the code so that, if the user does decide to change
the max number of Responses (for any Question type), the already
existing setMaxResponses() function is called instead of doing it
in-place in modifyQuestion().