Modification of Matching/Ranking; fixes to mods
The modification of Matching and Ranking has been completely finished.
Prompt can be changed, choices can be changed, and more choices can be
added. If more choices are added, automatically gets the new correct
Response from the user, in a Test. Fixed how ShortAnswerResponse() gets
the maxLength_ attribute of the ShortAnswer class; it is now passed
through a Set<String> using the getValidResponses() method. (The Essay
now also passes -1 through the same method, maintaining the old, hackish
behavior.) This allowed for the removal of the extra parameter for
constructor of ShortAnsweResponse as well.