Fixed incorrect Test grading; program finalized
A semicolon after the condition of an if statement was giving incorrect
grades. It was hard to notice because the first time I tested I was
simply trying to give the user a perfect score to test Response's
.equals(). After a massive amount of debugging, though, I caught the
semicolon at the end of the line. With that,  this version of the
program is finalized.