Made the InputOutput member variables static
This limits the program to only one type of I/O, but, as of right now,
there is only one type of I/O being used. It could theoretically be
fixed later, if a mixing of I/O types were needed. (For example, if some
questions were to have audio output, others console, and others video;
for now, the program is fine though.)